Day 6 part 2

+++ src/main/java/com/jacoblucas/adventofcode2023/day06/Race.java
@@ -1,4 +1,4 @@
 package com.jacoblucas.adventofcode2023.day06;
 
-public record Race(int id, int time, int recordDistance) {
+public record Race(int id, long time, long recordDistance) {
 }
